Back-up from HD, unrevised.
The files sent have a basic implementation of the map-making and walking simulator included in Walker-Prototype. In order to start the fight you need to achieve the X mark in the map using the O movable "character".

Therefore, the game included is a base for a turn-based RPG, focusing in Pokemon, but meant to be modular both in implementation and values edition.